Добавлен функционал получения информации о текущем пользователе, логаута. Все операции с контактами доступны только авторизированым пользователям и пользователям-владельцам контактов

+++ model/contacts.js
@@ -1,41 +1,42 @@
 const Contact = require("./schemas/contact");
 
-const listContacts = async () => {
+const listContacts = async (userID) => {
   try {
-    return await Contact.find({});
+    console.log(`list contacts user ID = ${userID}`);
+    return await Contact.find({ owner: userID });
   } catch (error) {
     console.log(error.message);
   }
 };
 
-const getContactById = async (id) => {
+const getContactById = async (userID, id) => {
   try {
-    return await Contact.findOne({ _id: id });
+    return await Contact.findOne({ _id: id, owner: userID });
   } catch (error) {
     console.log(error.message);
   }
 };
 
-const removeContact = async (id) => {
+const removeContact = async (userID, id) => {
   try {
-    return await Contact.findByIdAndRemove({ _id: id });
+    return await Contact.findByIdAndRemove({ _id: id, owner: userID });
   } catch (error) {
     console.log(error.message);
   }
 };
 
-const addContact = async (body) => {
+const addContact = async (userID, body) => {
   try {
-    return await Contact.create(body);
+    return await Contact.create({ ...body, owner: userID });
   } catch (error) {
     console.log(error.message);
   }
 };
 
-const updateContact = async (id, body) => {
+const updateContact = async (userID, id, body) => {
   try {
     return await Contact.findOneAndUpdate(
-      { _id: id },
+      { _id: id, owner: userID },
       { ...body },
       { new: true }
     );
@@ -44,10 +45,10 @@ const updateContact = async (id, body) => {
   }
 };
 
-const updateStatusContact = async (id, body) => {
+const updateStatusContact = async (userID, id, body) => {
   try {
     const result = await Contact.findOneAndUpdate(
-      { _id: id },
+      { _id: id, owner: userID },
       { ...body },
       { new: true }
     );
